+++ backend/main/resources/rating.py
@@ -3,28 +3,44 @@ from flask import request,jsonify
 from main.models import (RatingModel,UserModel,ProductModel)
 from sqlalchemy import func, desc
 from .. import db
+from flask_jwt_extended import jwt_required, get_jwt_identity, get_jwt
+from main.auth.decorators import role_required
 
 class Rating(Resource):
+    @jwt_required()
     def get(self,rating_id):
         rating = db.session.query(RatingModel).get_or_404(rating_id)
         return rating.to_json()
     
+    @role_required(roles = ["Admin","Users"])
     def delete(self,rating_id):
+        role = get_jwt().get('role')
         rating = db.session.query(RatingModel).get_or_404(rating_id)
+        if role == 'Users' and rating.user_id != get_jwt_identity():
+            return 'No tiene permisos para eliminar este recurso', 403
+        
         db.session.delete(rating)
         db.session.commit()
         return rating.to_json(), 200  
     
+ 
+    @role_required(roles = ["Users"])
     def put(self,rating_id):
         rating = db.session.query(RatingModel).get_or_404(rating_id)
-        data = request.get_json().items()
-        for key, value in data:
-            setattr(rating, key, value)
-        db.session.add(rating)
-        db.session.commit()
-        return rating.to_json(), 201
-    
+        current_identity = get_jwt_identity()
+        if current_identity == rating.user_id:
+
+            data = request.get_json().items()
+            for key, value in data:
+                setattr(rating, key, value)
+            db.session.add(rating)
+            db.session.commit()
+            return rating.to_json(), 201
+        else: 
+            return 'No tienes permisos para modificar este rating', 403
+
 class RatingList(Resource):
+    @role_required(roles = ["Admin"])
     def get(self):
 
         page =1 
@@ -60,7 +76,7 @@ class RatingList(Resource):
                         'total':rating.total,
                         'pages':rating.pages,
                         'page':page})
-    
+    @jwt_required()
     def post(self):
         rating = RatingModel.from_json(request.get_json())
         db.session.add(rating)
